+++ Source/GameFiles/Items/Inventory.h
@@ -21,7 +21,7 @@ public:
 	Inventory(const unsigned int inventorySize);
 	virtual ~Inventory();
 	/*! Search for the Item in and returns its position*/
-	int Search(const int ID) const;
+	int Search(const long ID) const;
 	/*! Adds an item to the array*/
 	bool AddItem(Item* item);
 	/*! Returns the item at the position */
@@ -29,7 +29,7 @@ public:
 	/*! Returns the Items*/
 	std::vector<Item*> GetItems() const;
 	/*! Search for the item and returns it if found or NULL if not found*/
-	Item* SearchAndGetItem(const unsigned int ID) const;
+	Item* SearchAndGetItem(const long ID) const;
 	/*! Removes the item*/
 	bool RemoveItem(const unsigned int index);
 	bool RemoveItem(Item* item);
@@ -42,9 +42,9 @@ public:
 	/*! Returns amount of Items the player has in the inventory with itemType = TYPE*/
 	int SearchForItemType(const unsigned int TYPE);
 	/*! Removes a stack from Item*/
-	bool RemoveItemStack(const int ID, const unsigned int numberOfStacks = 1);
+	bool RemoveItemStack(const long ID, const unsigned int numberOfStacks = 1);
 	Item* SearchAndGetItemFromType(const unsigned int TYPE);
-	Item* EquipItem(const int ID);
+	Item* EquipItem(const long ID);
 private:
 	/*! Array of items*/
 	std::vector<Item*> zItems;
