+++ Source/GameFiles/Items/MeleeWeapon.cpp
@@ -1,6 +1,6 @@
 #include "MeleeWeapon.h"
 
-MeleeWeapon::MeleeWeapon(const unsigned int id, const unsigned int itemType, const float damage, const float range)
+MeleeWeapon::MeleeWeapon(const long id, const unsigned int itemType, const float damage, const float range)
 	: Weapon(id, itemType, damage, range)
 {
 	this->zStacking = false;
