+++ Source/GameFiles/Items/MeleeWeapon.h
@@ -10,7 +10,7 @@
 class MeleeWeapon : public Weapon
 {
 public:
-	MeleeWeapon(const unsigned int id, const unsigned int itemType, const float damage, const float range);
+	MeleeWeapon(const long id, const unsigned int itemType, const float damage, const float range);
 	virtual ~MeleeWeapon();
 
 	virtual bool Use();
