+++ Source/GameFiles/Items/RangedWeapon.cpp
@@ -1,6 +1,6 @@
 #include "RangedWeapon.h"
 
-RangedWeapon::RangedWeapon(const unsigned int id, const unsigned int itemType, const float damage, const float range)
+RangedWeapon::RangedWeapon(const long id, const unsigned int itemType, const float damage, const float range)
 						   : Weapon(id, itemType, damage, range)
 {
 	this->zStacking = false;
