+++ Source/GameFiles/Items/RangedWeapon.h
@@ -10,7 +10,7 @@
 class RangedWeapon : public Weapon
 {
 public:
-	RangedWeapon(const unsigned int id, const unsigned int itemType, const float damage, const float range);
+	RangedWeapon(const long id, const unsigned int itemType, const float damage, const float range);
 	virtual ~RangedWeapon();
 
 	virtual bool Use();
